+++ api-derive/balances/account.ts
@@ -2,6 +2,7 @@
 import type { Observable } from 'https://esm.sh/rxjs@7.8.1';
 import type { QueryableStorageEntry } from 'https://deno.land/x/polkadot/api-base/types/index.ts';
 import type { AccountData, AccountId, AccountIndex, AccountInfo, Address, Balance, Index } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
+import type { FrameSystemAccountInfo, PalletBalancesAccountData } from 'https://deno.land/x/polkadot/types/lookup.ts';
 import type { ITuple } from 'https://deno.land/x/polkadot/types/types/index.ts';
 import type { DeriveApi, DeriveBalancesAccount, DeriveBalancesAccountData } from '../types.ts';
 
@@ -13,7 +14,9 @@ import { memo } from '../util/index.ts';
 
 type BalanceResult = [Balance, Balance, Balance, Balance];
 
-type Result = [Index, BalanceResult[]];
+type Result = [Index, BalanceResult[], AccountType];
+
+interface AccountType { isFrameAccountData: boolean }
 
 type DeriveCustomAccount = DeriveApi['derive'] & Record<string, {
   customAccount?: DeriveApi['query']['balances']['account']
@@ -23,24 +26,38 @@ function zeroBalance (api: DeriveApi) {
   return api.registry.createType('Balance');
 }
 
-function getBalance (api: DeriveApi, [freeBalance, reservedBalance, frozenFee, frozenMisc]: BalanceResult): DeriveBalancesAccountData {
+function getBalance (api: DeriveApi, [freeBalance, reservedBalance, frozenFeeOrFrozen, frozenMiscOrFlags]: BalanceResult, accType: AccountType): DeriveBalancesAccountData {
   const votingBalance = api.registry.createType('Balance', freeBalance.toBn());
+
+  if (accType.isFrameAccountData) {
+    return {
+      frameSystemAccountInfo: {
+        flags: frozenMiscOrFlags,
+        frozen: frozenFeeOrFrozen
+      },
+      freeBalance,
+      frozenFee: api.registry.createType('Balance', 0),
+      frozenMisc: api.registry.createType('Balance', 0),
+      reservedBalance,
+      votingBalance
+    };
+  }
 
   return {
     freeBalance,
-    frozenFee,
-    frozenMisc,
+    frozenFee: frozenFeeOrFrozen,
+    frozenMisc: frozenMiscOrFlags,
     reservedBalance,
     votingBalance
   };
 }
 
-function calcBalances (api: DeriveApi, [accountId, [accountNonce, [primary, ...additional]]]: [AccountId, Result]): DeriveBalancesAccount {
+function calcBalances (api: DeriveApi, [accountId, [accountNonce, [primary, ...additional], accType]]: [AccountId, Result]): DeriveBalancesAccount {
   return objectSpread({
     accountId,
     accountNonce,
-    additional: additional.map((b) => getBalance(api, b))
-  }, getBalance(api, primary));
+    additional: additional.map((b) => getBalance(api, b, accType))
+  }, getBalance(api, primary, accType));
 }
 
 function queryBalancesFree (api: DeriveApi, accountId: AccountId): Observable<Result> {
@@ -51,7 +68,8 @@ function queryBalancesFree (api: DeriveApi, accountId: AccountId): Observable<Re
   ]).pipe(
     map(([freeBalance, reservedBalance, accountNonce]): Result => [
       accountNonce,
-      [[freeBalance, reservedBalance, zeroBalance(api), zeroBalance(api)]]
+      [[freeBalance, reservedBalance, zeroBalance(api), zeroBalance(api)]],
+      { isFrameAccountData: false }
     ])
   );
 }
@@ -59,7 +77,8 @@ function queryBalancesFree (api: DeriveApi, accountId: AccountId): Observable<Re
 function queryNonceOnly (api: DeriveApi, accountId: AccountId): Observable<Result> {
   const fill = (nonce: Index): Result => [
     nonce,
-    [[zeroBalance(api), zeroBalance(api), zeroBalance(api), zeroBalance(api)]]
+    [[zeroBalance(api), zeroBalance(api), zeroBalance(api), zeroBalance(api)]],
+    { isFrameAccountData: false }
   ];
 
   return isFunction(api.query.system.account)
@@ -80,7 +99,8 @@ function queryBalancesAccount (api: DeriveApi, accountId: AccountId, modules: st
 
   const extract = (nonce: Index, data: AccountData[]): Result => [
     nonce,
-    data.map(({ feeFrozen, free, miscFrozen, reserved }): BalanceResult => [free, reserved, feeFrozen, miscFrozen])
+    data.map(({ feeFrozen, free, miscFrozen, reserved }): BalanceResult => [free, reserved, feeFrozen, miscFrozen]),
+    { isFrameAccountData: false }
   ];
 
   // NOTE this is for the first case where we do have instances specified
@@ -103,7 +123,7 @@ function queryBalancesAccount (api: DeriveApi, accountId: AccountId, modules: st
 
 function querySystemAccount (api: DeriveApi, accountId: AccountId): Observable<Result> {
   // AccountInfo is current, support old, eg. Edgeware
-  return api.query.system.account<AccountInfo | ITuple<[Index, AccountData]>>(accountId).pipe(
+  return api.query.system.account<AccountInfo | FrameSystemAccountInfo | ITuple<[Index, AccountData]>>(accountId).pipe(
     map((infoOrTuple): Result => {
       const data = (infoOrTuple as AccountInfo).nonce
         ? (infoOrTuple as AccountInfo).data
@@ -114,16 +134,30 @@ function querySystemAccount (api: DeriveApi, accountId: AccountId): Observable<R
       if (!data || data.isEmpty) {
         return [
           nonce,
-          [[zeroBalance(api), zeroBalance(api), zeroBalance(api), zeroBalance(api)]]
+          [[zeroBalance(api), zeroBalance(api), zeroBalance(api), zeroBalance(api)]],
+          { isFrameAccountData: false }
         ];
       }
 
-      const { feeFrozen, free, miscFrozen, reserved } = data;
+      const isFrameType = !!(infoOrTuple as FrameSystemAccountInfo).data.frozen;
 
-      return [
-        nonce,
-        [[free, reserved, feeFrozen, miscFrozen]]
-      ];
+      if (isFrameType) {
+        const { flags, free, frozen, reserved } = (data as unknown as PalletBalancesAccountData);
+
+        return [
+          nonce,
+          [[free, reserved, frozen, flags]],
+          { isFrameAccountData: true }
+        ];
+      } else {
+        const { feeFrozen, free, miscFrozen, reserved } = data;
+
+        return [
+          nonce,
+          [[free, reserved, feeFrozen, miscFrozen]],
+          { isFrameAccountData: false }
+        ];
+      }
     })
   );
 }
@@ -165,7 +199,8 @@ export function account (instanceId: string, api: DeriveApi): (address: AccountI
           ])
           : of([api.registry.createType('AccountId'), [
             api.registry.createType('Index'),
-            [[zeroBalance(api), zeroBalance(api), zeroBalance(api), zeroBalance(api)]]
+            [[zeroBalance(api), zeroBalance(api), zeroBalance(api), zeroBalance(api)]],
+            { isFrameAccountData: false }
           ]])
         )
       ),